+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Update translations (19.10.2022)
 - Tray icon is now by default enabled. Settings got moved to Settings->Advanced
 - instantly react to changing chat background color
+- made frontend code more independent from electron
 - reduce notifications when many messages are received at once.
 
 ### Fixed